requirements.txt
pydantic>=2.6,<3
python-dotenv>=1.0
pytest>=8.0
google-genai>=1.0
feedparser>=6.0
pyyaml>=6.0
